+++ bm/GhostWebView.swift
@@ -123,8 +123,8 @@ class GhostWebView: WKWebView {
 
     convenience init() {
         let webConfiguration = WKWebViewConfiguration()
-        // self.init(frame: CGRect(x: 0, y: 0, width: 300, height: 300), configuration: webConfiguration)
-        self.init(frame: CGRect(x: 0, y: 0, width: 1, height: 1), configuration: webConfiguration)
+        // self.init(frame: CGRect(x: 0, y: 0, width: 600, height: 300), configuration: webConfiguration)
+        self.init(frame: CGRect(x: 0, y: 0, width: 600, height: 1), configuration: webConfiguration)
     }
 
     func loadGhostPage() {