+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - [x] Restore web search
 - [x] Restore loans library info
 - [x] Improve loans parsing error message
-- [x] Improve loans parsing reliability 
+- [x] Improve loans parsing reliability
 
 
 ## 1.8